# internal/api/swagger.yaml
openapi: 3.0.3
info:
  title: Blog API
  description: |
    REST API for the brutalist blog CMS.

    **Block `content` encoding by type:**
    | type | content value |
    |------|--------------|
    | `text` | plain string |
    | `blockquote` | plain string |
    | `image` | JSON string: `{"src":"url","caption":"text"}` |
    | `audio` | JSON string: `{"label":"FILENAME.WAV","caption":"text"}` — also set `original_audio_url` |
    | `gallery` | JSON string: `[{"src":"url","alt":"text"},...]` |
    | `list` | JSON string: `["item 1","item 2","item 3"]` |
  version: 1.0.0
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Obtain a token from POST /admin/login and pass it as: Authorization: Bearer <token>"
  schemas:
...
    DashboardStats:
      type: object
      properties:
        posts:
          type: object
          properties:
            total:     { type: integer, example: 10 }
            published: { type: integer, example: 7 }
            draft:     { type: integer, example: 3 }
        blocks_by_type:
          type: array
          items:
            type: object
            properties:
              type:  { type: string, example: text }
              count: { type: integer, example: 20 }
        views:
          type: object
          properties:
            total:        { type: integer, example: 1234 }
            last_7_days:  { type: integer, example: 89 }
            last_30_days: { type: integer, example: 456 }
        top_posts:
          type: array
          items:
            type: object
            properties:
              id:         { type: integer, example: 1 }
              title:      { type: string, example: "ARCHITECTURAL VOID" }
              slug:       { type: string, example: "architectural-void" }
              view_count: { type: integer, example: 234 }
    Error:
      type: object
      properties:
        error: { type: string, example: "Invalid email or password" }
    Message:
      type: object
      properties:
        message: { type: string, example: "Deleted successfully" }
...
paths:
...
  /api/posts/{id}/view:
    post:
      tags: [Analytics]
      summary: Record a post view
      description: Call this from the frontend whenever a reader opens a post page. Pass the post slug as `{id}`. No authentication required.
      parameters:
        - name: id
          in: path
          required: true
          description: The post slug
          schema: { type: string, example: "architectural-void-and-digital-ephemerality" }
      responses:
        '200': { description: View recorded, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '404': { description: Post not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/admin/analytics:
    get:
      tags: [Analytics]
      summary: Get admin dashboard stats
      description: Returns post counts by status, block counts by type, view totals, and the top 5 most-viewed posts.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardStats' }
              example:
                posts:
                  total: 10
                  published: 7
                  draft: 3
                blocks_by_type:
                  - { type: text, count: 20 }
                  - { type: image, count: 8 }
                  - { type: audio, count: 4 }
                views:
                  total: 1234
                  last_7_days: 89
                  last_30_days: 456
                top_posts:
                  - { id: 1, title: "ARCHITECTURAL VOID", slug: "architectural-void", view_count: 234 }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '500': { description: Server error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }